Move utilities to Common.
