+++ Common/Utilities/Extensions/NPCExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using Terraria.GameContent;
 
-namespace PathOfTerraria.Helpers.Extensions;
+namespace PathOfTerraria.Common.Utilities.Extensions;
 
 public static class NPCExtensions
 {
