+++ Common/Utilities/StringUtils.cs
@@ -1,7 +1,7 @@
 ﻿using ReLogic.Graphics;
 using System.Collections.Generic;
 
-namespace PathOfTerraria.Helpers;
+namespace PathOfTerraria.Common.Utilities;
 
 public static class StringUtils
 {
@@ -43,7 +43,7 @@ public static class StringUtils
 				// If the next character is a CJK punctuation, we add both characters as a single word
 				// Unless the next character is a right close CJK punctuation (e.g. left brackets), in which case we add only the current character
 				if (i + 1 < input.Length && LocalizationUtils.IsCjkPunctuation(input[i + 1]) &&
-				    !LocalizationUtils.IsRightCloseCjkPunctuation(input[i + 1]))
+					!LocalizationUtils.IsRightCloseCjkPunctuation(input[i + 1]))
 				{
 					words.Add(input[i].ToString() + input[i + 1]);
 					i++;
@@ -97,7 +97,7 @@ public static class StringUtils
 
 		return output;
 	}
-	
+
 	/// <summary>
 	/// Gets the size of a texture - The mod name is automatically prepended to the texture path.
 	/// </summary>
